+++ alerts/src/main/java/com/thousandeyes/sdk/alerts/model/LegacyAlert.java
@@ -55,7 +55,7 @@ public class LegacyAlert {
   private String dateEnd;
 
   public static final String JSON_PROPERTY_RULE_ID = "ruleId";
-  private Integer ruleId;
+  private Long ruleId;
 
   /**
    * Current state of the alert. Possible values: clear or trigger.
@@ -155,7 +155,7 @@ public class LegacyAlert {
     @JsonProperty(JSON_PROPERTY_ALERT_ID) UUID alertId, 
     @JsonProperty(JSON_PROPERTY_DATE_START) String dateStart, 
     @JsonProperty(JSON_PROPERTY_DATE_END) String dateEnd, 
-    @JsonProperty(JSON_PROPERTY_RULE_ID) Integer ruleId, 
+    @JsonProperty(JSON_PROPERTY_RULE_ID) Long ruleId, 
     @JsonProperty(JSON_PROPERTY_STATE) StateEnum state
   ) {
   this();
@@ -219,7 +219,7 @@ public class LegacyAlert {
   @JsonProperty(JSON_PROPERTY_RULE_ID)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
 
-  public Integer getRuleId() {
+  public Long getRuleId() {
     return ruleId;
   }
